Rules syntax redux (#264)
* WIP on rules-syntax redux

* don't proxy symbols on the path

this fixes a bug where some other code was looking for a symbol (toOpaqueRef) that shouldn't always exist, but the proxy made it always exist

* rules should update with the this.update(self syntax

---------
