+++ typescript/packages/lookslike-high-level/src/recipes/commentbox.tsx
@@ -1,13 +1,14 @@
 import {
-  recipe,
   derive,
-  stream,
   type OpaqueRef,
   UI,
   ifElse,
+  select,
+  Spell,
+  doc,
+  $
 } from "@commontools/common-builder";
 import { h } from "@commontools/common-html";
-import { cell } from "@commontools/common-runner";
 
 /**
  * Plan to rewrite this on top of recipe stack
@@ -20,7 +21,7 @@ import { cell } from "@commontools/common-runner";
  * .compile() will call `recipe` with a callback that constructs all the opaque
  * refs and so on.
  *
- * E.g. this.set({ ... }) should be possible to do, we look up the cells
+ * E.g. this.update({ ... }) should be possible to do, we look up the cells
  * in our cell table and update them.
  *
  * A few changes:
@@ -36,221 +37,9 @@ import { cell } from "@commontools/common-runner";
  *   - `match(condition)` adds a condition to the rule
  *   - First step: Just assume `select`.
  * - handlers just get values
- * - event handlers use this.set() without self to update the state
+ * - event handlers use this.update() without self to update the state
  * - derive rules just return new values
  */
-
-// $ is a proxy that just collect paths, so that one can call [getPath] on it
-// and get an array. For example for `q = $.foo.bar[0]` `q[getPath]` yields
-// `["foo", "bar", 0]`. This is used to generate queries.
-
-type PathSegment = PropertyKey | { fn: string; args: any[] };
-const getPath = Symbol("getPath");
-
-// Create the path collector proxy
-function createPathCollector(path: PathSegment[] = []): any {
-  return new Proxy(
-    function () {}, // Base target is a function to support function calls
-    {
-      get(_target, prop) {
-        if (prop === getPath) return path;
-        // Continue collecting path for property access
-        return createPathCollector([...path, prop]);
-      },
-
-      // Catch any function calls
-      apply(_target, _thisArg, args) {
-        // Get the last segment which should be the function name
-        const lastSegment = path[path.length - 1];
-        if (typeof lastSegment !== "string") {
-          throw new Error("Invalid function call");
-        }
-
-        // Remove the function name from the path and add it as a function call
-        const newPath = path.slice(0, -1);
-        return createPathCollector([...newPath, { fn: lastSegment, args }]);
-      },
-    },
-  );
-}
-
-// Create the root $ proxy
-const $ = createPathCollector();
-
-// Resolve $ to a paths on `self`
-// TODO: Also for non-top-level ones
-function resolve$(self: OpaqueRef<any>, query: any) {
-  const entries = Object.entries(query);
-  const result: Record<string, any> = {};
-
-  for (const [key, value] of entries) {
-    if (value && typeof value === "object" && getPath in value) {
-      const path = value[getPath] as PathSegment[];
-
-      let current = self;
-      for (const segment of path) {
-        if (typeof segment === "object" && "fn" in segment) {
-          // Execute any function with its arguments
-          current = current[segment.fn].apply(current, segment.args);
-        } else {
-          current = current[segment as PropertyKey];
-        }
-      }
-
-      result[key] = current;
-    } else {
-      result[key] = value;
-    }
-  }
-
-  return result;
-}
-
-function select(query: any) {
-  const generateQuery = (self: OpaqueRef<any>) => resolve$(self, query);
-  Object.assign(generateQuery, {
-    with: (schema: any) =>
-      select({
-        ...resolve$(self, query),
-        ...Object.fromEntries(
-          Object.keys(schema.properties ?? {}).map(key => [
-            key,
-            self[key as any],
-          ]),
-        ),
-      }),
-  });
-  return generateQuery;
-}
-
-// addRule(event("update"), ({ $event }) => { ... })
-function event(name: string) {
-  // .compile() will replace $event with actual stream
-  return select({ $event: name });
-}
-
-export abstract class Spell<T extends Record<string, any>> {
-  private eventListeners: Array<{
-    type: string;
-    handlerFn: (self: any, ev: any) => any;
-  }> = [];
-  private rules: Array<{
-    condition: any;
-    handlerFn: (ctx: any) => any;
-  }> = [];
-
-  private streams: Record<string, OpaqueRef<any>> = {};
-
-  constructor() {}
-
-  // `self` is what is passed to the handler, so for now a query result proxy
-  set(self: any, values: Partial<T>) {
-    Object.entries(values).forEach(([key, value]) => {
-      self[key] = value;
-    });
-  }
-
-  // Is this being called when an event happens?
-  /*
-  dispatch(self: any, event: string, detail: any) {
-    return [
-      { Upsert: [self, appendOnPrefix(event), detail] }
-    ]
-  }
-  */
-
-  // Use this in JSX, e.g. onClick={this.dispatch('random')}
-  // TODO: Add details bound to event
-  dispatch(event: string) {
-    console.log("dispatch", event, this.streams[event]);
-    return this.streams[event];
-  }
-
-  addEventListener(type: string, handlerFn: (self: any, ev: any) => any) {
-    this.eventListeners.push({ type, handlerFn });
-  }
-
-  addRule(condition: any, handlerFn: (ctx: any) => any) {
-    this.rules.push({ condition, handlerFn });
-  }
-
-  abstract init(): T;
-
-  abstract render(state: T): any;
-
-  // Used when chaining the query, e.g. `with(this.get('meta', ""))`
-  /*
-  get<S extends string>(field: S, defaultValue?: any) {
-    if (defaultValue) {
-      return select({ [field]: $[field] } as const).clause(defaultTo($.self, field, $[field], defaultValue));
-    }
-    return select({ [field]: $[field] } as const).match($.self, field, $[field])
-  }
-  */
-
-  compile(title: string = "Spell") {
-    return recipe(title, (self: OpaqueRef<any>) => {
-      const initialState = this.init() ?? {};
-      const state: Record<string, OpaqueRef<any>> = {};
-
-      Object.entries(initialState).forEach(([key, value]) => {
-        self[key].setDefault(value);
-        state[key] = self[key];
-      });
-
-      this.eventListeners.forEach(({ type, handlerFn }) => {
-        this.streams[type] ??= stream();
-        derive({ self, $event: this.streams[type] }, ({ self, $event }) =>
-          handlerFn(self, $event),
-        );
-      });
-
-      this.rules.forEach(rule => {
-        // condition:
-        //  ($) => { foo: $.foo }
-        //  select({ foo: $.foo })
-        //  select({ foo: $.foo, bar: $.bar.map(item => item.foo) })
-        //     .filter(fn), count(), take(n), skip(n),
-        //     .sortBy(fn, dir?), groupBy(key), distinct(key),
-        //     .join(ref, key?)
-        //  event("update")
-        //  ["foo", "bar"]
-
-        let condition = rule.condition(self);
-
-        if (Array.isArray(condition))
-          condition = Object.fromEntries(
-            condition.map(key => [key, self[key]]),
-          );
-        else if (
-          condition &&
-          typeof condition === "object" &&
-          condition !== null &&
-          "$event" in condition
-        )
-          condition["$event"] = this.streams[condition["$event"]];
-
-        derive(condition, rule.handlerFn);
-      });
-
-      return {
-        [UI]: this.render(self),
-        ...this.streams,
-        ...state,
-      };
-    });
-  }
-}
-
-function doc<T = any>(value: any) {
-  return cell<T>(value).getAsQueryResult();
-}
-
-type CommentsState = {
-  title: string;
-  description: string;
-  meta: Meta | null;
-};
 
 type Meta = {
   category: string;
@@ -262,8 +51,7 @@ class MetadataSpell extends Spell<Meta> {
     super();
 
     this.addEventListener("update", (self, ev) => {
-      console.log("update", ev);
-      this.set(
+      this.update(
         self,
         doc({
           category: ev.detail.value,
@@ -303,6 +91,13 @@ function Metadata({ meta }: { meta: Meta }, _children: any) {
   return metadata(meta ?? {})[UI];
 }
 
+type CommentsState = {
+  title: string;
+  description: string;
+  meta: Meta | null;
+  metaLength: number;
+};
+
 export class CommentBoxSpell extends Spell<CommentsState> {
   constructor() {
     super();
@@ -315,12 +110,18 @@ export class CommentBoxSpell extends Spell<CommentsState> {
         submittedAt: new Date().toISOString(),
       });
 
-      this.set(self, { meta: metadata });
+      this.update(self, { meta: metadata });
     });
 
     this.addEventListener("submit", self => {
       console.log("submit", self);
     });
+
+    this.addRule(
+      select({ category: $.meta.category, title: $.title }),
+      ({ self, category, title }) =>
+        this.update(self, { metaLength: title.length + (category ?? "").length })
+    );
   }
 
   override init() {
@@ -328,14 +129,16 @@ export class CommentBoxSpell extends Spell<CommentsState> {
       title: "Class Syntax Demo",
       description: "It has an embedded doc!",
       meta: null,
+      metaLength: 0,
     };
   }
 
-  override render({ description, title, meta }: OpaqueRef<CommentsState>) {
+  override render({ description, title, meta, metaLength }: OpaqueRef<CommentsState>) {
     return (
       <div>
         <h1>{title}</h1>
         <p>{description}</p>
+        <p>metaLength: {metaLength}</p>
         <pre>{derive(meta, meta => JSON.stringify(meta, null, 2))}</pre>
         {ifElse(
           meta,
@@ -351,31 +154,8 @@ export class CommentBoxSpell extends Spell<CommentsState> {
       </div>
     );
   }
-
-  /*
-  override render({ description, title, meta }: CommentsState) {
-    return (
-      <div>
-        <h1>{title}</h1>
-        <p>{description}</p>
-        <pre>{JSON.stringify(meta, null, 2)}</pre>
-        {meta && (
-          <div>
-            <Metadata meta={{ category: "test", submittedAt: "test" }} />
-          </div>
-        )}
-        <common-form onsubmit={this.dispatch("submit")}>
-          <common-input type="text" name="message" />
-        </common-form>
-        <common-button onclick={this.dispatch("random")}>Random</common-button>
-      </div>
-    );
-  }*/
 }
 
 const commentBox = new CommentBoxSpell().compile("Comment Box");
-
-(window as any).metadata = metadata;
-(window as any).commentBox = commentBox;
 
 export default commentBox;
